refactor: webChessController 객체 만들고 메서드 이동
